+++ traits/TSingletone.php
@@ -1,0 +1,22 @@
+<?php
+
+declare(strict_types=1);
+
+namespace app\traits;
+
+trait TSingletone
+{
+    private static $instance = null;
+
+    private function __construct() {}
+    private function __clone() {}
+    private function __wakeup() {}
+
+    public static function getInstance()
+    {
+        if (is_null(static::$instance)) {
+            static::$instance = new static();
+        }
+        return static::$instance;
+    }
+}
